Moved function to util module
Updated plot_antennas_range_time
